moved common paginator logic to trait
